Added view session checking and new stock trigger.

+++ application/views/emr/emr_view.php
@@ -14,11 +14,16 @@
 
                 $this->table->set_heading('RecordID','DoctorID','RegisterID','Time of Admission','');
                 foreach($query as $row){
-                    $edit = "<a href='".site_url()."/emr/viewdata/".$row->record_id."' title='".$row->register_id."'>View</a>
-                    <br>
-                    <a href='".site_url()."/emr/editdata/".$row->record_id."' title='".$row->register_id."'>Edit</a>
-                    <br>
-                    <a href='".site_url()."/emr/deletedata/".$row->record_id."' title='".$row->register_id."' onclick='return confirmDelete();'>Delete</a>"; 
+                  if($_SESSION['status'] == "ADMIN"){
+                     $edit = "<a href='".site_url()."/emr/viewdata/".$row->record_id."' title='".$row->register_id."'>View</a>
+                      <br>
+                      <a href='".site_url()."/emr/editdata/".$row->record_id."' title='".$row->register_id."'>Edit</a>
+                      <br>
+                      <a href='".site_url()."/emr/deletedata/".$row->record_id."' title='".$row->register_id."' onclick='return confirmDelete();'>Delete</a>"; 
+                  } else {
+                    $edit = "<a href='".site_url()."/emr/viewdata/".$row->record_id."' title='".$row->register_id."'>View</a>"; 
+                  }
+                   
                     $this->table->add_row($row->record_id, $row->doctor_id, $row->register_id, $row->time ,$edit);
                 }
                 echo $this->table->generate();
