+++ application/views/payment/payment_view.php
@@ -14,10 +14,17 @@
 
                 $this->table->set_heading('PaymentID','RegisterID','WorkerID','Type of Payment','Amount','Time of Payment','');
                 foreach($query as $row){
-                    $edit = "<a href='".site_url()."/athenaReport/get_entry/".$row->register_id."' title='".$row->register_id."'>Print Entry</a>
-                    <br>
-                    <a href='".site_url()."/athenaReport/get_invoice/".$row->payment_id."' title='".$row->payment_id."'>Print</a><br>
-                    <a href='".site_url()."/payment/deletedata/".$row->payment_id."' title='".$row->register_id."' onclick='return confirmDelete();'>Delete</a>"; 
+                    if($_SESSION['status'] == "ADMIN"){
+                      $edit = "<a href='".site_url()."/athenaReport/get_entry/".$row->register_id."' title='".$row->register_id."'>Print Entry</a>
+                      <br>
+                      <a href='".site_url()."/athenaReport/get_invoice/".$row->payment_id."' title='".$row->payment_id."'>Print</a><br>
+                      <a href='".site_url()."/payment/deletedata/".$row->payment_id."' title='".$row->register_id."' onclick='return confirmDelete();'>Delete</a>"; 
+                    } else {
+                      $edit = "<a href='".site_url()."/athenaReport/get_entry/".$row->register_id."' title='".$row->register_id."'>Print Entry</a>
+                      <br>
+                      <a href='".site_url()."/athenaReport/get_invoice/".$row->payment_id."' title='".$row->payment_id."'>Print</a>"; 
+                    }
+                    
                     $this->table->add_row($row->payment_id, $row->register_id, $row->worker_id, $row->type, $row->amount, $row->time,$edit);
                 }
                 echo $this->table->generate();
